+++ app/database/database.go
@@ -31,13 +31,3 @@ func DBConnect() *gorm.DB {
 
 	return db
 }
-
-func GetUserData(Id int) (User, error) {
-	db := DBConnect()
-
-	user := User{}
-
-	err := db.Debug().Where("id = ?", Id).First(&user).Error
-
-	return user, err
-}
